backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
langgraph
openai
tavily-python
pydantic
python-dotenv
